final fix

+++ examples/a_plus_b.py
@@ -1,9 +1,9 @@
-from hackrform import read, display
+from hackerforms import read, display
 
 print("aqui tem print!")
 
-a = read("Qual o valor de a?", dtype=int)
-b = read("Qual o valor de b?", dtype=int)
+a = read("Qual o valor de a?")
+b = read("Qual o valor de b?")
 soma = a + b
 
 display("a + b = " + str(soma))
